Update HTML and CSS for Liam Johnson Ceramics website
This  includes updates to the HTML and CSS for the Liam Johnson Ceramics website.The navigation links have been given a class of "links". In the CSS, new styles have been added for the h1 and .links classes to display them inline. The .links li class has also been updated to display inline. The .links a class has been updated with new styles for color, text-decoration, font-size, and margin.