Add product cards to Liam Johnson Ceramics website
This  includes the addition of product cards to the HTML for the Liam Johnson Ceramics website. Each card contains an image and a h2 title.

In the CSS, new styles have been added for the .card, .card img, and .card h2 classes. The .card class has a width of 49.75%, is displayed inline, and has padding of 16px. The .card img class has a height of 500px and a width of 100%. The .card h2 class is displayed inline.